pie_app/requirements.txt
RPi.GPIO
picamera

flask
flask-cors
pyserial